test(setup): extract the console.error/warn trap into a shared module
Both the client (jsdom) and the new server (node) setups need the same
'unexpected console output fails the test' trap, so factor it out of dom.ts
into test/setup/console-trap.ts.
